requirements.txt
fastapi
uvicorn
streamlit
openai
python-dotenv
faiss-cpu
tiktoken
pydantic
numpy
pypdf
